Move ResetPassword command and handler into Core
