+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_resources/AdminResetPassword.xml
@@ -15,7 +15,7 @@
            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xsi:schemaLocation="https://api-platform.com/schema/metadata https://api-platform.com/schema/metadata/metadata-2.0.xsd"
 >
-    <resource class="Sylius\Bundle\ApiBundle\Command\Admin\ResetPassword" shortName="AdminResetPasswordRequest">
+    <resource class="Sylius\Bundle\CoreBundle\Message\Admin\Account\ResetPassword" shortName="AdminResetPasswordRequest">
         <attribute name="validation_groups">sylius</attribute>
         <attribute name="messenger">input</attribute>
 
@@ -39,7 +39,7 @@
             <itemOperation name="admin_update_reset_password_request">
                 <attribute name="method">PATCH</attribute>
                 <attribute name="path">/admin/reset-password-requests/{resetPasswordToken}</attribute>
-                <attribute name="input">Sylius\Bundle\ApiBundle\Command\Admin\ResetPassword</attribute>
+                <attribute name="input">Sylius\Bundle\CoreBundle\Message\Admin\Account\ResetPassword</attribute>
                 <attribute name="output">false</attribute>
                 <attribute name="status">202</attribute>
                 <attribute name="denormalization_context">
