+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/command_handlers.xml
@@ -188,12 +188,6 @@
             <tag name="messenger.message_handler" bus="sylius_default.bus" />
         </service>
 
-        <service id="Sylius\Bundle\ApiBundle\CommandHandler\Admin\ResetPasswordHandler">
-            <argument type="service" id="sylius.user_password_resetter.admin" />
-            <tag name="messenger.message_handler" bus="sylius.command_bus" />
-            <tag name="messenger.message_handler" bus="sylius_default.bus" />
-        </service>
-
         <service id="Sylius\Bundle\ApiBundle\CommandHandler\Checkout\SendOrderConfirmationHandler">
             <argument type="service" id="sylius.email_sender" />
             <argument type="service" id="sylius.repository.order" />
