+++ src/Sylius/Bundle/CoreBundle/Message/Admin/Account/ResetPassword.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\Command\Admin;
+namespace Sylius\Bundle\CoreBundle\Message\Admin\Account;
 
 /** @experimental */
 class ResetPassword
