+++ src/Sylius/Bundle/CoreBundle/Resources/config/serialization/Messages/Admin/Account/ResetPassword.xml
@@ -15,7 +15,7 @@
             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
             xsi:schemaLocation="http://symfony.com/schema/dic/serializer-mapping https://symfony.com/schema/dic/serializer-mapping/serializer-mapping-1.0.xsd"
 >
-    <class name="Sylius\Bundle\ApiBundle\Command\Admin\ResetPassword">
+    <class name="Sylius\Bundle\CoreBundle\Message\Admin\Account\ResetPassword">
         <attribute name="newPassword">
             <group>admin:reset_password:update</group>
         </attribute>
