+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/message_handlers.xml
@@ -28,6 +28,13 @@
             <tag name="messenger.message_handler" bus="sylius_default.bus" />
         </service>
 
+        <service id="Sylius\Bundle\CoreBundle\MessageHandler\Admin\Account\ResetPasswordHandler">
+            <argument type="service" id="sylius.user_password_resetter.admin" />
+
+            <tag name="messenger.message_handler" bus="sylius.command_bus" />
+            <tag name="messenger.message_handler" bus="sylius_default.bus" />
+        </service>
+
         <service id="Sylius\Bundle\CoreBundle\MessageHandler\Admin\Account\SendResetPasswordEmailHandler">
             <argument type="service" id="sylius.repository.admin_user"/>
             <argument type="service" id="sylius.email_sender"/>
